# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SkillRapido - Real-time service marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /providers/profile:
    get:
      operationId: getProviderProfile
      tags: [providers]
      summary: Get current provider's profile
      responses:
        "200":
          description: Provider profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProviderProfile"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    put:
      operationId: updateProviderProfile
      tags: [providers]
      summary: Update provider profile and skills
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProviderProfileBody"
      responses:
        "200":
          description: Updated provider profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProviderProfile"
  /providers/availability:
    put:
      operationId: updateProviderAvailability
      tags: [providers]
      summary: Toggle provider online/offline status
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AvailabilityBody"
      responses:
        "200":
          description: Updated availability
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProviderProfile"
  /providers/location:
    put:
      operationId: updateProviderLocation
      tags: [providers]
      summary: Update provider's current location
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LocationBody"
      responses:
        "200":
          description: Location updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /providers/jobs:
    get:
      operationId: getProviderJobs
      tags: [providers]
      summary: Get jobs assigned to the provider
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [active, completed, all]
          description: Filter by status
      responses:
        "200":
          description: Provider's jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Job"

  # JOBS
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    Skill:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
        icon:
          type: string
        description:
          type: string
          nullable: true
      required: [id, name, category, icon]
    ProviderProfile:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
          nullable: true
        bio:
          type: string
          nullable: true
        isOnline:
          type: boolean
        isVerified:
          type: boolean
        rating:
          type: number
          nullable: true
        totalJobs:
          type: integer
        totalEarnings:
          type: number
        skills:
          type: array
          items:
            $ref: "#/components/schemas/Skill"
        latitude:
          type: number
          nullable: true
        longitude:
          type: number
          nullable: true
      required: [id, userId, name, email, isOnline, isVerified, totalJobs, totalEarnings, skills]
    UpdateProviderProfileBody:
      type: object
      properties:
        bio:
          type: string
        phone:
          type: string
        skillIds:
          type: array
          items:
            type: integer
        latitude:
          type: number
        longitude:
          type: number
    AvailabilityBody:
      type: object
      properties:
        isOnline:
          type: boolean
      required: [isOnline]
    LocationBody:
      type: object
      properties:
        latitude:
          type: number
        longitude:
          type: number
      required: [latitude, longitude]
...
    Job:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [open, assigned, in_progress, completed, cancelled]
        budget:
          type: number
        commission:
          type: number
        providerAmount:
          type: number
        address:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        scheduledAt:
          type: string
          format: date-time
          nullable: true
        completedAt:
          type: string
          format: date-time
          nullable: true
        customerId:
          type: integer
        customerName:
          type: string
        providerId:
          type: integer
          nullable: true
        providerName:
          type: string
          nullable: true
        skill:
          $ref: "#/components/schemas/Skill"
        bidCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, title, description, status, budget, commission, providerAmount, address, latitude, longitude, customerId, customerName, skill, bidCount, createdAt]
...